requirements.txt
flask==3.0.0
DateTime==5.2
Flask-Caching==2.1.0
Flask-Session>=0.3.2,<0.6
flask-sqlalchemy>=3.1.1
flask-migrate>=4.0.5
mysql-connector-python>=8.0.33
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
werkzeug==3.0.1
click==8.1.7
Jinja2==3.1.2
jsonify==0.5
msal==1.24.1
ms_identity_web
pyopenssl>19,<20
pyodbc==5.2.0
